Revert "Fixes #527"
This reverts commit 07372bad2164fa8b4f0f55fc042f65a42ecdd4f6.

+++ konig-schemagen/src/main/java/io/konig/schemagen/gcp/BigQueryDatatype.java
@@ -26,8 +26,7 @@ public enum BigQueryDatatype {
 	STRING,
 	BYTES,
 	INTEGER,
-	//FLOAT,
-	FLOAT64,
+	FLOAT,
 	BOOLEAN,
 	TIMESTAMP,
 	RECORD
